requirements.txt
Flask==0.10.1
Flask-Login==0.2.11
Flask-RESTful==0.2.12
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5
Flask-WTF==0.10.2
Jinja2==2.7.3
LogentriesLogger==0.2.1
Mako==1.0.0
MarkupSafe==0.23
MySQL-python==1.2.5
SQLAlchemy==0.9.7
WTForms==2.0.1
Werkzeug==0.9.6
alembic==0.6.7
aniso8601==0.83
argparse==1.2.1
coverage==3.7.1
flake8==2.2.3
itsdangerous==0.24
mccabe==0.2.1
pep8==1.5.7
py==1.4.24
pyflakes==0.8.1
pytest==2.6.2
pytz==2014.7
six==1.8.0
wsgiref==0.1.2
